Refactor KNX component: Remove TP-UART data link layer and transport layer
- Deleted `tpuart_data_link_layer.h` and `transport_layer.cpp`/`transport_layer.h` files to streamline the KNX component.
- Added `.gitignore` for ESPHome specific files.
- Created `test.yaml` configuration for the KNX router using ESP32C6 with UART settings.

+++ esphome/components/knx/esphome_platform.cpp
@@ -53,7 +53,7 @@ size_t ESPHomePlatform::readBytesUart(uint8_t *buffer, size_t length) {
   return 0;
 }
 
-uint8_t *ESPHomePlatform::getEepromBuffer(uint16_t size) {
+uint8_t *ESPHomePlatform::getEepromBuffer(uint32_t size) {
   if (size > KNX_FLASH_SIZE) {
     return nullptr;
   }
